Add functionality for policy group tabs
This commit does 2 things:

1. Moves the nav item helpers that used to live in the editions_helper
to the tabbed_nav_helper.
2. Adds functionality and tests for when a group policy is passed in
instead of an edition.

There are further methods in this helper which deal with when we
port those sections on the application over in this helper such as
topics and people.
