+++ app/helpers/admin/tabbed_nav_helper.rb
@@ -1,4 +1,77 @@
 module Admin::TabbedNavHelper
+  include Admin::EditionsHelper
+
+  def secondary_navigation_tabs_items(model, current_path)
+    if model.is_a?(Edition)
+      edition_nav_items(model, current_path)
+    else
+      send("#{model.class.model_name.param_key}_nav_items", model, current_path)
+    end
+  end
+
+  def edition_nav_items(edition, current_path)
+    nav_items = []
+    nav_items << standard_edition_nav_items(edition, current_path)
+    nav_items << consultation_nav_items(edition, current_path) if edition.persisted? && edition.is_a?(Consultation)
+    nav_items << document_collection_nav_items(edition, current_path) if edition.persisted? && edition.is_a?(DocumentCollection)
+    nav_items.flatten
+  end
+
+  def standard_edition_nav_items(edition, current_path)
+    [
+      {
+        label: "Document",
+        href: tab_url_for_edition(edition),
+        current: current_path == tab_url_for_edition(edition),
+      },
+      *(if edition.persisted? && edition.allows_attachments?
+          [{
+            label: "Attachments#{tag.span(edition.attachments.count, class: 'govuk-tag govuk-tag--grey') if edition.attachments.count.positive?}".html_safe,
+            href: admin_edition_attachments_path(edition),
+            current: current_path == admin_edition_attachments_path(edition),
+          }]
+        end),
+    ]
+  end
+
+  def consultation_nav_items(edition, current_path)
+    [
+      {
+        label: "Public feedback",
+        href: admin_consultation_public_feedback_path(edition),
+        current: current_path == admin_consultation_public_feedback_path(edition),
+      },
+      {
+        label: "Final outcome",
+        href: admin_consultation_outcome_path(edition),
+        current: current_path == admin_consultation_outcome_path(edition),
+      },
+    ]
+  end
+
+  def document_collection_nav_items(edition, current_path)
+    {
+      label: "Collection documents",
+      href: admin_document_collection_groups_path(edition),
+      current: current_path == admin_document_collection_groups_path(edition),
+    }
+  end
+
+  def policy_group_nav_items(group, current_path)
+    [
+      {
+        label: "Group",
+        href: edit_admin_policy_group_path(group),
+        current: current_path == edit_admin_policy_group_path(group),
+      },
+      {
+        label: "Attachments#{tag.span(group.attachments.count, class: 'govuk-tag govuk-tag--grey') if group.attachments.count.positive?}".html_safe,
+        href: admin_policy_group_attachments_path(group),
+        current: current_path == admin_policy_group_attachments_path(group),
+      },
+    ]
+  end
+
   def tab_navigation_for(model, *extra_classes, &block)
     tabs = send("#{model.class.model_name.param_key}_tabs", model)
     tab_navigation(tabs, *extra_classes, &block)
